Create banned_books_01
The House of the Spirits, the unforgettable first novel that established Isabel Allende as one of the world’s most gifted storytellers, brings to life the triumphs and tragedies of three generations of the Trueba family. The patriarch Esteban is a volatile, proud man whose voracious pursuit of political power is tempered only by his love for his delicate wife Clara, a woman with a mystical connection to the spirit world. When their daughter Blanca embarks on a forbidden love affair in defiance of her implacable father, the result is an unexpected gift to Esteban: his adored granddaughter Alba, a beautiful and strong-willed child who will lead her family and her country into a revolutionary future.